chore: rename chart.ts -> core.ts
