+++ src/core/core.ts
