+++ src/core/index.ts
@@ -1,2 +1,1 @@
-export { chart } from "./chart";
-export type { ChartInterface } from "./chart";
+export { Core } from "./core";
